added voronoi function
